Misleading attribute name
The parameter names of `getCellColor` function is not matching the invocations made to it in `renderProductTable`. I guess we either need to change the invocations to provide `index` as the second argument (since thats what contains the row number) like: `getCellColor(0, index)` or change the argument name to `column` (since that's what with the acutal invocations we pass in).